chore(build): replace go.tools.txt with go.tools.go and a build tag
This is so dependencies are marked as used and not cleanup up upon a `go
mod tidy`. Note: given this is now listing all imports from package `.`,
if any other file gets added to this root package, its modules will be
listed. The build flag only means it is not included in the standard
build process, but when using tags, they are _in addition_ to the non
tagged packages/files. Future improvement: find a better way to scope
this go tools list down without being too permisive.